Adjusted files to be of smaller size, moved main menu to main cpp file, general reorganization
