+++ salesMenu.h
@@ -1,0 +1,12 @@
+#pragma once //  compiler only sees this lib file once
+
+#include <iostream>
+#include <string>
+#include <vector>
+#include <cctype>
+#include <cstdlib>
+#include <stdexcept>
+#include "utils.h"
+
+void salesMenu(); //TODO
+    void productsMenu(); //TODO
